+++ R/index-by.R
@@ -70,9 +70,9 @@ index_by.tbl_ts <- function(.data, ...) {
   }
   expr_name <- names(exprs)[1]
   idx <- index(.data)
-  idx_chr <- quo_text(idx)
+  idx_chr <- as_string(idx)
   if (identical(idx_chr, expr_name)) {
-    abort(sprintf("LHS must not be named as index (`%s`).", idx_chr))
+    abort(sprintf("Column `%s` (index) can't be overwritten.", idx_chr))
   }
   # ungroup() protect the index class
   tbl <- mutate(ungroup(.data), !!! exprs)
@@ -86,14 +86,14 @@ index_by.tbl_ts <- function(.data, ...) {
 
 index_rename <- function(.data, .vars) {
   names <- names(.vars)
-  idx_chr <- quo_text(index(.data))
+  idx_chr <- as_string(index(.data))
   new_idx_chr <- names[idx_chr == .vars]
   sym(new_idx_chr)
 }
 
 index2_rename <- function(.data, .vars) {
   names <- names(.vars)
-  idx_chr <- quo_text(index2(.data))
+  idx_chr <- as_string(index2(.data))
   new_idx_chr <- names[idx_chr == .vars]
   sym(new_idx_chr)
 }
@@ -105,71 +105,4 @@ index2_update <- function(.data, .vars) {
   } else {
     sym(chr)
   }
-}
-
-tsibble_rename <- function(.data, ...) {
-  names_dat <- names(.data)
-  val_vars <- tidyselect::vars_rename(names_dat, ...)
-
-  # index
-  idx <- index_rename(.data, val_vars)
-  # index2
-  idx2 <- index2_rename(.data, val_vars)
-  attr(.data, "index2") <- idx2
-  # key (key of the same size (bf & af))
-  new_key <- key_rename(.data, val_vars)
-  # groups
-  new_grp <- grp_rename(.data, val_vars)
-  attr(.data, "vars") <- new_grp
-
-  names(.data) <- names(val_vars)
-  build_tsibble(
-    .data, key = new_key, index = !! idx, index2 = !! idx2,
-    groups = new_grp, regular = is_regular(.data), validate = FALSE, 
-    ordered = is_ordered(.data), interval = interval(.data)
-  )
-}
-
-tsibble_select <- function(.data, ..., validate = TRUE) {
-  dots <- c(enquos(...), new_quosure(index(.data)))
-  names_dat <- names(.data)
-  val_vars <- tidyselect::vars_select(names_dat, !!! dots)
-  sel_data <- select(as_tibble(.data), !!! val_vars)
-
-  # checking
-  # val_idx <- has_index(j = val_vars, x = .data)
-  # if (is_false(val_idx)) {
-  #   abort(sprintf(
-  #     "The `index` (`%s`) must not be dropped, do you want `.drop = TRUE` to drop `tbl_ts`?",
-  #     quo_text(index(.data))
-  #   ))
-  # }
-  
-  # index
-  idx <- index_rename(.data, val_vars)
-  # index2
-  idx2 <- index2_rename(.data, val_vars)
-  # key (key of the reduced size (bf & af) but also different names)
-  key_vars <- val_vars[val_vars %in% key_vars(.data)]
-  tmp_data <- key_remove(.data, key_vars, validate = FALSE)
-  new_key <- key_rename(tmp_data, key_vars)
-  # groups
-  new_grp <- grp_rename(.data, val_vars)
-  
-  if (validate) {
-    vec_names <- union(names(val_vars), names(.data))
-    # either key or index is present in ...
-    # suggests that the operations are done on these variables
-    # validate = TRUE to check if tsibble still holds
-    # val_idx <- has_index(vec_names, .data)
-    # val_key <- has_any_key(vec_names, .data)
-    # validate <- val_idx || val_key
-    validate <- has_any_key(vec_names, .data)
-  }
-  
-  build_tsibble(
-    sel_data, key = new_key, index = !! idx, index2 = !! idx2,
-    groups = new_grp, regular = is_regular(.data), validate = validate, 
-    ordered = is_ordered(.data), interval = interval(.data)
-  )
 }